Do not exit lock screen when Power is pressed
We should not exit lock screen for any reason.

Closes: #84

+++ compositor/qml/desktop/ScreenView.qml
@@ -285,7 +285,8 @@ Rectangle {
         case Qt.Key_PowerDown:
         case Qt.Key_Suspend:
         case Qt.Key_Hibernate:
-            screenView.state = "poweroff";
+            if (screenView.state != "lock")
+                screenView.state = "poweroff";
             event.accepted = true;
             return;
         default:
